# Conflicts:
#	app/src/androidTest/java/ch/epfl/favo/view/FavorPageTest.java
#	app/src/main/java/ch/epfl/favo/MainActivity.java
#	app/src/main/java/ch/epfl/favo/view/tabs/FavorPage.java
#	app/src/main/java/ch/epfl/favo/view/tabs/addFavor/FavorDetailView.java
#	app/src/main/java/ch/epfl/favo/view/tabs/addFavor/FavorRequestView.java

+++ app/src/sharedTestUtils/java/ch/epfl/favo/FakeItemFactory.java
@@ -14,7 +14,7 @@ public class FakeItemFactory {
         TestConstants.DESCRIPTION,
         TestConstants.REQUESTER_ID,
         TestConstants.LOCATION,
-        TestConstants.FAVOR_STATUS);
+        TestConstants.FAVOR_STATUS.toInt());
   }
 
   public static List<Favor> getFavorList() {
